Correction float --> int pour la reconversion des notes

+++ Src/Prediction/g2_ratings_pred.py
@@ -8,4 +8,4 @@ def notes(df,id_user,n,id_item):
     M=df.progress_apply(axis=0,func= np.mean)
     tqdm.pandas(desc = "Ecarts types")
     E=df.progress_apply(axis=0,func= np.std)
-    return n*E[id_item]+M[id_item]
+    return int(n*E[id_item]+M[id_item])
